docs: Fix workspace index type mismatches and method call errors - resolve 28 compilation errors across input handling and IPC (TEAM_032)
- Fix 7 workspace index type conversions in src/input/mod.rs and src/input/move_grab.rs:
  - Cast ws_idx to usize for view_offset_gesture_begin() calls (lines 2596, 3074, 3561, 124)
- Fix 3 output reference comparisons in src/input/spatial_movement_grab.rs:
  - Change Some(&self.output) to Some(self.output) for comparison
  - Cast ws_idx to usize in view_offset_gesture_begin() call

+++ src/input/mod.rs
@@ -2593,7 +2593,7 @@ impl State {
                     };
                     self.niri
                         .layout
-                        .view_offset_gesture_begin(&output, Some(ws_idx), false);
+                        .view_offset_gesture_begin(&output, Some(ws_idx as usize), false);
                     let grab = SpatialMovementGrab::new(start_data, output, ws_id, true);
                     pointer.set_grab(self, grab, serial, Focus::Clear);
                     self.niri
@@ -3071,7 +3071,7 @@ impl State {
 
                                 self.niri.layout.view_offset_gesture_begin(
                                     &output,
-                                    Some(ws_idx),
+                                    Some(ws_idx as usize),
                                     true,
                                 );
                                 redraw = true;
@@ -3558,7 +3558,7 @@ impl State {
                             let ws_idx = self.niri.layout.find_workspace_by_id(ws.id()).unwrap().0;
                             self.niri
                                 .layout
-                                .view_offset_gesture_begin(&output, Some(ws_idx), true);
+                                .view_offset_gesture_begin(&output, Some(ws_idx as usize), true);
                         }
                     } else {
                         self.niri
